separate style from structure
